proxmox_reader/requirements.txt
proxmoxer
python-dotenv